requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes -o requirements.txt
amqp==5.3.1
    # via kombu
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
beautifulsoup4==4.14.3
    # via markdownify
billiard==4.2.4
    # via celery
celery==5.6.3
    # via knight
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   rich-toolkit
    #   typer
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
cryptography==46.0.7
    # via
    #   google-auth
    #   pyjwt
distro==1.9.0
    # via google-genai
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.135.3
    # via knight
fastapi-cli==0.0.24
    # via fastapi
fastapi-cloud-cli==0.17.0
    # via fastapi-cli
fastar==0.11.0
    # via fastapi-cloud-cli
filetype==1.2.0
    # via langchain-google-genai
google-auth==2.49.2
    # via google-genai
google-genai==1.73.1
    # via langchain-google-genai
greenlet==3.4.0 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   knight
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
kombu==5.6.2
    # via celery
langchain==1.2.15
    # via knight
langchain-core==1.2.26
    # via
    #   langchain
    #   langchain-google-genai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==4.2.1
    # via knight
langgraph==1.1.6
    # via
    #   knight
    #   langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.9
    # via langgraph
langgraph-sdk==0.3.12
    # via langgraph
langsmith==0.7.25
    # via langchain-core
markdown-it-py==4.0.0
    # via rich
markdownify==1.2.2
    # via knight
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   kombu
    #   langchain-core
    #   langsmith
prompt-toolkit==3.0.52
    # via click-repl
psycopg==3.3.3
    # via knight
psycopg-binary==3.3.3 ; implementation_name != 'pypy'
    # via psycopg
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   knight
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.1
    # via fastapi
pydantic-settings==2.13.1
    # via
    #   fastapi
    #   knight
pygments==2.20.0
    # via rich
pyjwt==2.12.1
    # via knight
pymysql==1.1.2
    # via knight
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.26
    # via fastapi
python-tds==1.17.1
    # via knight
pyyaml==6.0.3
    # via
    #   langchain-core
    #   uvicorn
redis==6.4.0
    # via kombu
requests==2.33.1
    # via
    #   google-auth
    #   google-genai
    #   knight
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==15.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.19.7
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.58.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
sniffio==1.3.1
    # via google-genai
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.49
    # via knight
starlette==1.0.0
    # via fastapi
tenacity==9.1.4
    # via
    #   google-genai
    #   langchain-core
typer==0.24.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   fastapi
    #   google-genai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via
    #   kombu
    #   psycopg
    #   tzlocal
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.44.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.6.0
    # via prompt-toolkit
websockets==16.0
    # via
    #   google-genai
    #   uvicorn
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zstandard==0.25.0
    # via langsmith
